+++ src/main/java/es/udc/fi/dc/fd/service/PermissionChecker.java
@@ -5,8 +5,8 @@ import es.udc.fi.dc.fd.model.persistence.UserImpl;
 
 public interface PermissionChecker {
 	
-	public void checkUserExists(int userId) throws InstanceNotFoundException;
+	public void checkUserExists(Long userId) throws InstanceNotFoundException;
 	
-	public UserImpl checkUser(int userId) throws InstanceNotFoundException;
+	public UserImpl checkUser(Long userId) throws InstanceNotFoundException;
 	
 }
